% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_aic.R
\name{summarize_aic}
\alias{summarize_aic}
\title{Return a tidy tibble of AIC information given a list of model objects}
\usage{
summarize_aic(models)
}
\arguments{
\item{models}{List of model objects}
}
\value{
Tibble of AIC summary stats
}
\description{
Return a tidy tibble of AIC information given a list of model objects
}
