% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_models.R
\name{summarize_models}
\alias{summarize_models}
\title{Print tidied tables of model summary stats}
\usage{
summarize_models(models, ...)
}
\arguments{
\item{models}{List of model objects}

\item{...}{args to be passed to `Anova`}
}
\description{
Print tidied tables of model summary stats
}
