% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_predictions.R
\name{tidy_predictions}
\alias{tidy_predictions}
\title{Build a single tibble of predictions from a list of models}
\usage{
tidy_predictions(models)
}
\arguments{
\item{models}{List of model objects}
}
\value{
Tibble of fitted predictions
}
\description{
Build a single tibble of predictions from a list of models
}
